single_RAG_agent_no_tool/requirements.txt
# Microsoft Agent Framework (preview)
agent-framework --pre
agent-framework-azure-ai --pre

# Azure SDKs
azure-identity

# Local vector store
chromadb

# Web scraping & text processing
requests
beautifulsoup4
lxml

# Utilities
python-dotenv
pydantic
